Move to single index.tsx for each parcel package
