+++ packages/parcel-optimizer/src/index.tsx
@@ -1,1 +1,8 @@
-export { default } from './optimizer';
+import { sort } from '@compiled/css';
+import { Optimizer } from '@parcel/plugin';
+
+export default new Optimizer({
+  async optimize({ contents, map }) {
+    return { contents: sort(contents.toString()), map };
+  },
+});
